requirements.txt
hikari
hikari-lightbulb
pyyaml
sqlalchemy
apscheduler
git+https://github.com/Rapptz/asqlite